pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "cryptography>=45.0.6",
    "flask-limiter>=3.12",
    "bleach>=6.2.0",
    "qrcode[pil]>=8.2",
    "pillow>=11.3.0",
]
